Move remove_tone to a separate file
