+++ underthesea/pipeline/sentiment/general/text_features.py
@@ -2,7 +2,7 @@ from sklearn.base import BaseEstimator, TransformerMixin
 import string
 import re
 from underthesea.pipeline.word_tokenize import tokenize
-from underthesea.utils.vietnamese_ipa import remove_tone
+from underthesea.utils.vietnamese_features import remove_tone
 
 negative_emoticons = {':(', '☹', '❌', '👎', '👹', '💀', '🔥', '🤔', '😏', '😐', '😑', '😒', '😓', '😔', '😕', '😖',
                       '😞', '😟', '😠', '😡', '😢', '😣', '😤', '😥', '😧', '😨', '😩', '😪', '😫', '😭', '😰', '😱',
